Mas refactorizacion de codigo.
