+++ tests/Field/Hint/ExceptionTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Forge\Form\Tests\Base\Field\Error;
+namespace Forge\Form\Tests\Field\Hint;
 
-use Forge\Form\Base\Field\Error;
 use Forge\Form\Exception\AttributeNotSet;
+use Forge\Form\Field\Hint;
 use Forge\Form\Tests\Support\PropertyTypeForm;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
@@ -20,7 +20,7 @@ final class ExceptionTest extends TestCase
     {
         $this->expectException(AttributeNotSet::class);
         $this->expectExceptionMessage('Failed to create widget because "attribute" is not set or not exists.');
-        Error::create(construct: [new PropertyTypeForm(), '']);
+        Hint::create(construct: [new PropertyTypeForm(), '']);
     }
 
     /**
@@ -30,6 +30,6 @@ final class ExceptionTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Tag name cannot be empty.');
-        Error::create(construct: [new PropertyTypeForm(), 'string'])->tag('')->render();
+        Hint::create(construct: [new PropertyTypeForm(), 'string'])->tag('')->render();
     }
 }
